refactor: 🔧 Jules05: Resolve cross-agent conflict — Risk Management & Feature Pipeline Harmonization
Consolidates duplicate risk management implementations (RiskManager vs RiskEngine)
into a single authoritative RiskManager supporting the 8-layer institutional cascade.
Relocates FeatureEngineer to the data domain and updates system-wide imports.
Ensures backward compatibility for legacy Kelly-sizing and audit paths.
