+++ main.py
@@ -27,7 +27,7 @@ if TYPE_CHECKING:
 
     from src.core.audit_log import AuditLogger
     from src.core.decision_support import DecisionSupportSystem
-    from src.core.feature_engineering import FeatureEngineer
+    from src.data.feature_engineering import FeatureEngineer
     from src.core.monitor import Monitor
     from src.core.schemas import TradeSignal
     from src.core.trade_logger import TradeLogger
@@ -619,11 +619,13 @@ def run_live(
                                     )
 
                                     # Update allocator performance for feedback loop
-                                    if updated_trade and allocator:
-                                        strat_id = f"{cfg.algorithm.upper()}_{cfg.symbol}_{cfg.timeframe}"
-                                        allocator.update_strategy_performance(
-                                            strat_id, updated_trade.pnl
-                                        )
+                                    if updated_trade:
+                                        risk.record_pnl(updated_trade.pnl)
+                                        if allocator:
+                                            strat_id = f"{cfg.algorithm.upper()}_{cfg.symbol}_{cfg.timeframe}"
+                                            allocator.update_strategy_performance(
+                                                strat_id, updated_trade.pnl
+                                            )
                             closed_tickets.append(symbol)
 
                     if closed_tickets and trade_logger:
@@ -1462,7 +1464,7 @@ def main() -> int:
             )
             return 1
     from src.core.decision_support import DecisionSupportSystem
-    from src.core.feature_engineering import FeatureEngineer
+    from src.data.feature_engineering import FeatureEngineer
     from src.core.health import HealthStatus, init_health_checker
     from src.core.trade_logger import TradeLogger
     from src.models.ensemble import EnsembleModel
