+++ src/trading/__init__.py
@@ -12,7 +12,7 @@ from src.trading.capital_allocator import (
     StrategyConfig,
 )
 from src.trading.mt5_connector import MT5Connector
-from src.trading.risk_engine import RiskDecision, RiskEngine
+from src.trading.risk_manager import RiskDecision, RiskManager
 
 __all__ = [
     "AllocationRequest",
@@ -23,7 +23,7 @@ __all__ = [
     "PerformanceReport",
     "RejectionCode",
     "RiskDecision",
-    "RiskEngine",
+    "RiskManager",
     "StrategyConfig",
     "TradeSignal",
 ]
