+++ src/trading/audited_risk_manager.py
@@ -9,11 +9,13 @@ License: MIT
 from __future__ import annotations
 
 import logging
-from typing import Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.audit_log import get_audit_logger
 from src.core.schemas import TradeSignal
-from src.trading.risk_manager import RiskManager
+from src.trading.risk_manager import RiskDecision, RiskManager
 
 logger = logging.getLogger(__name__)
 
@@ -24,6 +26,60 @@ class AuditedRiskManager(RiskManager):
     Evaluates the full decision chain for traceability.
     """
 
+    def validate_signal(
+        self,
+        signal: TradeSignal,
+        market_data: pd.DataFrame,
+        open_positions: List[Dict[str, Any]],
+        model_health: Optional[Dict[str, float]] = None,
+    ) -> RiskDecision:
+        """
+        Validate a trade signal and log the full decision chain to the audit log.
+        """
+        decision_chain = {
+            "circuit_breaker": self._check_drawdown_breaker(),
+            "daily_loss": self.get_daily_loss_level() < 4,
+            "max_trades": self.daily.trade_count < self.cfg.max_trades_per_day,
+            "max_positions": len(open_positions) < self.cfg.max_positions,
+            "directional_exposure": self._check_directional_exposure(signal, open_positions),
+            "total_notional": self._check_total_notional(signal, open_positions, market_data),
+            "symbol_allocation": self._check_symbol_allocation(signal.symbol),
+            "min_confidence": self._check_minimum_confidence(signal.confidence),
+            "risk_reward": self._check_risk_reward(signal),
+            "consecutive_losses": self._check_consecutive_losses(),
+            "model_health": self._check_model_health(model_health),
+        }
+
+        passed = all(decision_chain.values())
+        adjusted_lots = 0.0
+        reason = "Approved"
+
+        if passed:
+            adjusted_lots = self.size_position(signal.symbol, market_data=market_data)
+            if adjusted_lots < self.cfg.min_lot_size:
+                passed = False
+                reason = f"Calculated lot size {adjusted_lots} below minimum"
+                decision_chain["lot_size"] = False
+        else:
+            rejection_reasons = [k for k, v in decision_chain.items() if not v]
+            reason = f"Failed filters: {', '.join(rejection_reasons)}"
+
+        self._log_audit(signal, decision_chain, passed)
+
+        if not passed:
+            logger.warning(
+                "Signal REJECTED | %s %s | %s",
+                signal.symbol,
+                signal.direction,
+                reason,
+            )
+            if self.monitor:
+                rejection_reasons = [k for k, v in decision_chain.items() if not v]
+                for r in rejection_reasons:
+                    self.monitor.record_internal_rejection("risk_manager", r.upper())
+
+        return RiskDecision(is_approved=passed, reason=reason, adjusted_lot_size=adjusted_lots)
+
     def approve(
         self,
         signal: TradeSignal,
@@ -31,14 +87,12 @@ class AuditedRiskManager(RiskManager):
         model_health: Optional[dict] = None,
     ) -> bool:
         """
-        Run the full 8-layer risk filter cascade.
-        Returns True only if ALL layers pass.
-        Logs the full decision chain to the audit log.
+        Legacy compatibility wrapper for validate_signal with audit logging.
         """
         decision_chain = {
-            "circuit_breaker": self._check_circuit_breaker(),
-            "daily_loss": self._check_daily_loss(),
-            "max_positions": self._check_max_positions(),
+            "circuit_breaker": self._check_drawdown_breaker(),
+            "daily_loss": self.get_daily_loss_level() < 4,
+            "max_positions": len(self.open_positions) < self.cfg.max_positions,
             "symbol_allocation": self._check_symbol_allocation(signal.symbol),
             "min_confidence": self._check_minimum_confidence(signal.confidence),
             "risk_reward": self._check_risk_reward(signal),
@@ -48,35 +102,7 @@ class AuditedRiskManager(RiskManager):
 
         passed = all(decision_chain.values())
 
-        # Log to Audit Trail
-        try:
-            audit = get_audit_logger()
-            audit.log_risk_decision(
-                symbol=signal.symbol,
-                direction=signal.direction,
-                decision_chain=decision_chain,
-                passed=passed,
-            )
-
-            # Log high-severity circuit breaker events specifically
-            if not decision_chain.get("circuit_breaker", True):
-                audit.log_operator_action(
-                    operator="system",
-                    action="circuit_breaker_triggered",
-                    reason=f"Hard drawdown limit hit during signal validation for {signal.symbol}",
-                    metadata={"symbol": signal.symbol, "decision_chain": decision_chain},
-                )
-
-            if not decision_chain.get("daily_loss", True):
-                audit.log_operator_action(
-                    operator="system",
-                    action="daily_loss_limit_triggered",
-                    reason=f"Daily loss limit reached during signal validation for {signal.symbol}",
-                    metadata={"symbol": signal.symbol, "decision_chain": decision_chain},
-                )
-
-        except (RuntimeError, ImportError):
-            logger.debug("AuditLogger not available for risk decision logging")
+        self._log_audit(signal, decision_chain, passed)
 
         if not passed:
             rejection_reasons = [k for k, v in decision_chain.items() if not v]
@@ -98,3 +124,16 @@ class AuditedRiskManager(RiskManager):
                     signal_id=signal_id,
                 )
         return passed
+
+    def _log_audit(self, signal: TradeSignal, decision_chain: Dict[str, bool], passed: bool) -> None:
+        """Common audit logging logic."""
+        try:
+            audit = get_audit_logger()
+            audit.log_risk_decision(
+                symbol=signal.symbol,
+                direction=signal.direction.value,
+                decision_chain=decision_chain,
+                passed=passed,
+            )
+        except (RuntimeError, ImportError):
+            logger.debug("AuditLogger not available for risk decision logging")
