+++ tests/test_data_pipeline_integration.py
@@ -45,7 +45,7 @@ with patch.dict("sys.modules", {
 }):
     # Import components we want to test
     from src.core.constants import SignalDirection
-    from src.core.feature_engineering import FeatureEngineer
+    from src.data.feature_engineering import FeatureEngineer
     from src.models import ensemble  # Use the module to avoid attribute errors on reload
     from src.models.ensemble import EnsembleModel
     from src.utils.synthetic_data import ScenarioGenerator
@@ -104,9 +104,9 @@ def test_data_to_model_inference_flow(data_generator, feature_engineer, mock_ens
     df = data_generator.generate(n_steps=n_steps, regime="trending")
     df.index = pd.date_range(start="2024-01-01", periods=n_steps, freq="1min")
 
-    # The FeatureEngineer uses the 'talib' module imported at the top of src/core/feature_engineering.py
+    # The FeatureEngineer uses the 'talib' module imported at the top of src.data.feature_engineering.py
     # We must patch it there.
-    import src.core.feature_engineering as fe_mod
+    import src.data.feature_engineering as fe_mod
     with patch.object(fe_mod, "talib", mock_talib):
         setup_mock_talib(mock_talib)
         features = feature_engineer.compute_features(df)
@@ -130,7 +130,7 @@ def test_pipeline_resilience_to_malformed_data(data_generator, feature_engineer,
     df_malformed = data_generator.generate(n_steps=100, regime="malformed")
     df_malformed.index = pd.date_range(start="2024-01-01", periods=100, freq="1min")
 
-    import src.core.feature_engineering as fe_mod
+    import src.data.feature_engineering as fe_mod
     with patch.object(fe_mod, "talib", mock_talib):
         setup_mock_talib(mock_talib)
         features = feature_engineer.compute_features(df_malformed)
@@ -146,7 +146,7 @@ def test_pipeline_insufficient_history(data_generator, feature_engineer):
     df_short = data_generator.generate(n_steps=n_short, regime="ranging")
     df_short.index = pd.date_range(start="2024-01-01", periods=n_short, freq="1min")
 
-    import src.core.feature_engineering as fe_mod
+    import src.data.feature_engineering as fe_mod
     with patch.object(fe_mod, "talib", mock_talib):
         setup_mock_talib(mock_talib)
         # Force NaNs to simulate indicator warmup
@@ -162,7 +162,7 @@ def test_mtf_feature_alignment(data_generator, feature_engineer):
     df = data_generator.generate(n_steps=n_steps, regime="ranging")
     df.index = pd.date_range(start="2024-01-01", periods=n_steps, freq="1min")
 
-    import src.core.feature_engineering as fe_mod
+    import src.data.feature_engineering as fe_mod
     with patch.object(fe_mod, "talib", mock_talib):
         setup_mock_talib(mock_talib)
         features = feature_engineer.compute_features(df)
